refactor(filters): Split out html filters
